third_party/blink/web_tests/external/wpt/html/dom/reflection-metadata-expected.txt

This is a testharness.js-based test.
Found 52 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] link.as: setAttribute() to "audio"
  assert_equals: IDL get expected "audio" but got ""
[FAIL] link.as: setAttribute() to "AUDIO"
  assert_equals: IDL get expected "audio" but got ""
[FAIL] link.as: setAttribute() to "document"
  assert_equals: IDL get expected "document" but got ""
[FAIL] link.as: setAttribute() to "DOCUMENT"
  assert_equals: IDL get expected "document" but got ""
[FAIL] link.as: setAttribute() to "embed"
  assert_equals: IDL get expected "embed" but got ""
[FAIL] link.as: setAttribute() to "EMBED"
  assert_equals: IDL get expected "embed" but got ""
[FAIL] link.as: setAttribute() to "manifest"
  assert_equals: IDL get expected "manifest" but got ""
[FAIL] link.as: setAttribute() to "MANIFEST"
  assert_equals: IDL get expected "manifest" but got ""
[FAIL] link.as: setAttribute() to "object"
  assert_equals: IDL get expected "object" but got ""
[FAIL] link.as: setAttribute() to "OBJECT"
  assert_equals: IDL get expected "object" but got ""
[FAIL] link.as: setAttribute() to "report"
  assert_equals: IDL get expected "report" but got ""
[FAIL] link.as: setAttribute() to "REPORT"
  assert_equals: IDL get expected "report" but got ""
[FAIL] link.as: setAttribute() to "sharedworker"
  assert_equals: IDL get expected "sharedworker" but got ""
[FAIL] link.as: setAttribute() to "SHAREDWORKER"
  assert_equals: IDL get expected "sharedworker" but got ""
[FAIL] link.as: setAttribute() to "video"
  assert_equals: IDL get expected "video" but got ""
[FAIL] link.as: setAttribute() to "VIDEO"
  assert_equals: IDL get expected "video" but got ""
[FAIL] link.as: setAttribute() to "worker"
  assert_equals: IDL get expected "worker" but got ""
[FAIL] link.as: setAttribute() to "WORKER"
  assert_equals: IDL get expected "worker" but got ""
[FAIL] link.as: setAttribute() to "xslt"
  assert_equals: IDL get expected "xslt" but got ""
[FAIL] link.as: setAttribute() to "XSLT"
  assert_equals: IDL get expected "xslt" but got ""
[FAIL] link.as: IDL set to "audio"
  assert_equals: IDL get expected "audio" but got ""
[FAIL] link.as: IDL set to "AUDIO"
  assert_equals: IDL get expected "audio" but got ""
[FAIL] link.as: IDL set to "document"
  assert_equals: IDL get expected "document" but got ""
[FAIL] link.as: IDL set to "DOCUMENT"
  assert_equals: IDL get expected "document" but got ""
[FAIL] link.as: IDL set to "embed"
  assert_equals: IDL get expected "embed" but got ""
[FAIL] link.as: IDL set to "EMBED"
  assert_equals: IDL get expected "embed" but got ""
[FAIL] link.as: IDL set to "manifest"
  assert_equals: IDL get expected "manifest" but got ""
[FAIL] link.as: IDL set to "MANIFEST"
  assert_equals: IDL get expected "manifest" but got ""
[FAIL] link.as: IDL set to "object"
  assert_equals: IDL get expected "object" but got ""
[FAIL] link.as: IDL set to "OBJECT"
  assert_equals: IDL get expected "object" but got ""
[FAIL] link.as: IDL set to "report"
  assert_equals: IDL get expected "report" but got ""
[FAIL] link.as: IDL set to "REPORT"
  assert_equals: IDL get expected "report" but got ""
[FAIL] link.as: IDL set to "sharedworker"
  assert_equals: IDL get expected "sharedworker" but got ""
[FAIL] link.as: IDL set to "SHAREDWORKER"
  assert_equals: IDL get expected "sharedworker" but got ""
[FAIL] link.as: IDL set to "video"
  assert_equals: IDL get expected "video" but got ""
[FAIL] link.as: IDL set to "VIDEO"
  assert_equals: IDL get expected "video" but got ""
[FAIL] link.as: IDL set to "worker"
  assert_equals: IDL get expected "worker" but got ""
[FAIL] link.as: IDL set to "WORKER"
  assert_equals: IDL get expected "worker" but got ""
[FAIL] link.as: IDL set to "xslt"
  assert_equals: IDL get expected "xslt" but got ""
[FAIL] link.as: IDL set to "XSLT"
  assert_equals: IDL get expected "xslt" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
Harness: the test ran to completion.
